统一玩家状态类型为int32

+++ common/mjoption/xingpai.go
@@ -11,14 +11,14 @@ import (
 
 // XingPaiOption 行牌选项
 type XingPaiOption struct {
-	ID                   int    `yaml:"id"`                      // 选项 ID
-	WallCards            []int  `yaml:"wall_cards"`              // 墙牌
-	HuGameOver           bool   `yaml:"hu_gameover"`             // 胡牌后是否触发游戏结束
-	Hnz                  Hnz    `yaml:"hnz"`                     // 是否开启换N张
-	EnableDingque        bool   `yaml:"enable_dingque"`          // 是否开启定缺
-	EnableAddflower      bool   `yaml:"enable_addflower"`        // 是否开启补花
-	EnableChi            bool   `yaml:"enable_chi"`              // 是否可以吃牌
-	PlayerNoNormalStates uint64 `yaml:"player_no_normal_states"` // 玩家不参与游戏的不正常状态
+	ID                   int   `yaml:"id"`                      // 选项 ID
+	WallCards            []int `yaml:"wall_cards"`              // 墙牌
+	HuGameOver           bool  `yaml:"hu_gameover"`             // 胡牌后是否触发游戏结束
+	Hnz                  Hnz   `yaml:"hnz"`                     // 是否开启换N张
+	EnableDingque        bool  `yaml:"enable_dingque"`          // 是否开启定缺
+	EnableAddflower      bool  `yaml:"enable_addflower"`        // 是否开启补花
+	EnableChi            bool  `yaml:"enable_chi"`              // 是否可以吃牌
+	PlayerNoNormalStates int32 `yaml:"player_no_normal_states"` // 玩家不参与游戏的不正常状态
 	// Zhuang        int            `yaml:"zhuang"`              // 选庄的方式
 }
